Remove gulp in favor of npm scripts
Closes #45

Also, this addresses the issue many had with the static includes that
were generated.  Many thanks to @tubaman for bring this issue to light
in #30 and while this commit makes his PR unnecessary to merge, it was
his work in that PR and the follow on discussion that included @brosner
pointing out that Django now includes a static storage backend that
will handle the cache busting hash naming of static files for us.

Closes #30